更新服务商注册

+++ ananops-provider/ananops-provider-uac/src/main/java/com/ananops/provider/service/UacGroupService.java
@@ -99,7 +99,7 @@ public interface UacGroupService extends IService<UacGroup> {
 	 *
 	 * @return the int
 	 */
-	int saveUacGroup(UacGroup group, LoginAuthDto loginAuthDto);
+	Long saveUacGroup(UacGroup group, LoginAuthDto loginAuthDto);
 
 	/**
 	 * Gets by id.
